Don't pass invalid encodings to DBF. (#43)
* Don't pass invalid encodings to DBF.

When a CPG file is present, we read it to determine the encoding of the DBF file. Unfortunately, this file can specify encodings that Ruby doesn't actually know about (e.g. ANSI 1252). In this case, we should fallback to letting the DBF gem determine the encoding.

Closes https://github.com/rgeo/rgeo-shapefile/issues/39

* Handle trailing newlines in the CPG file

When the CPG file is manually edited, some text editors will automatically add a trailing newline. We should handle this.